Move Trillian implemention out of the way
I have come to the conclusion that Trillian is not the way to go.
It's too heavy to operate and not suitable for the Nix community as a
foundation for something like Trustix.

We need far lighter weight options, especially considering
mobile-nixos where we might want to runt Trustix on phones, even as
log operators.