decider: Add an aggregated decider wrapping multiple underlying ones
This is meant to support the functionality of having more than one
decider listed and stepping through the engines one by one, so that
another engine is used as a fallback in case the previous one didn't match.

You could easily imagine a user wanting to to a best-effort
verification, falling back to only verifying cache.nixos.org if it's
the only log available.