meta: Unpin nixpkgs
For now we're unpinning nixpkgs and relying on the developers NIX_PATH
to point to unstable.

This is done so that it's easier for me to do quick nixpkgs fixups in
development instead of always having to generate patches and apply on
top of the currently pinned nixpkgs.

Once all NixOS modules and E2E tests are done we'll revisit this and
use a ready-made tool for pinning instead of custom scripting.