contrib/ui-dashboard: Don't pull in all of Hydra, just the evaluator
Ideally we would split the evaluator from the rest of the Hydra source
code but for now let's just prevent having the rest of the tools from
hydra in $PATH.

I'm working on the split in a separate repository since this is very
generally applicable and the Hydra evaluator is interesting for many
more use cases than just Trustix.

However, it is not ready and requires a well thought out design, so
for now we will stick with the evaluator from Hydra proper until such
a time arises.