contrib/ui-dashboard: Add frontend templates with classical backend template rendering
The "modern" JS-heavy approach turned out to be too cumbersome to
develop and the Trustix dashis a fairly simple thing.

We are better off using traditional methods that are much simpler and
put less strain on the clients and developers alike (we don't have
that many frontend developers in the Nix community).

We're still gonna need some JS on the frontend but let's make the
scope of that as small as possible.

An additional bonus of this is that I think it's much easier to make
an accessible website when the content is largely text based instead
of dynamically created on the client.