packages/trustix: Change client connections to be "pooled" by log id
This is the final piece of the puzzle to make Trustix possible to federate.
We still don't have any code to do so in practice but with this change
it's architecturally possible.

This works by decoupling the what (the log) from the how (the
connection) and allows any connection to serve up any log, all this
without actually changing any trust properties.

This also removes custom TLS verification