packages/trustix: Switch DB from badger to bbolt
Badger has some nasty implications such as an inherent max transaction
size which caps out at a bit more than 500 entries.

We want to support much bigger transactions than that, so we need to
switch the database.

The big downside to this is lower performance but that shouldn't
really matter much for this use case.