packages/trustix: Add JSON config file support
JSON and TOML are semantically compatible so compatibility wise this
is not an issue.

The motivation for this is that we don't really expect users to write
configs by hand but use NixOS modules and such so it's kind of
silly that the NixOS module have to go `Nix -> JSON -> TOML`.
With this change we can skip that extra transpilation step and a
full derivation per system rebuild.