Also index fixed output derivations
I don't think it's a good idea to filter them out.
It's probably better to allow looking them up via the API if only to
be able to tell a querying user that it's impossible to get a hash
mismatch and also not have random holes in the dependency graph.